moved sampling related functions in Scene to under path namespace
